requirements.txt
asgiref==3.3.4
boto==2.49.0
boto3==1.17.69
botocore==1.20.69
certifi==2020.12.5
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
wincertstore==0.2
